Improve radio function, remove combokey instead activate radio wheel on tilde key.
This was a tough fix: as the only way to open the radio wheel was to use SetControlNormal which emulates 'Q' being pressed. Due to thread speed discrepancies IsControlJustReleased(0, tilde) did not work.

I check if the key is pressed if so while it is pressed set a local variable radio_wheel_active and add check in main LVC thread. Then apply a delay after release to ensure the lights do not get toggled on radio wheel exit (Q release).