Feature: support for wildcard vehicle profiles
`#` is a wildcard for 1 or more digits.

Example: UpdateApprovedTones will attempt to find
SIREN_ASSIGNEMENTS['veh1'] if not found then it will try and find a wildcard match
SIREN_ASSIGNEMENTS['veh#'] if not found then revert to
SIREN_ASSIGNEMENTS['DEFAULT'].

It is important to note `#` is for 1 OR MORE digits. Meaning SIREN_ASSIGNEMENTS['veh1'], SIREN_ASSIGNEMENTS['veh12'], SIREN_ASSIGNEMENTS['veh123'], etc. will all equate to wildcard 'veh#'.

The use of multiple '#' next to each other will not work, and is unnecessary.